Send_Hexadecimal: handle more then 50 chars and improve error reporting
The latest Send_Hexadecimal changes introduce a hard limit of 50 chars, which
is undesirable, this patch fixes this.

While at it, I noticed that Send_Hexadecimal skips parts of the string which
it cannot parse. This patch improvers error handling by sending nothing when
encountering input it cannot parse, and alerting the user.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>