Extend the probing space for serial ports; check for validity
On Linux, the entire namespace /dev/tty[A-Z]* is available for serial
ports, so probe it. To make it easier to modify namespaces in the
future, use the glib regular expression interface.

Use readdir() to probe instead of explicitly opening sequential
numbers.

At least on Linux, some drivers -- notably the 8250 UART driver --
leaves a lot of non-existent entries in /dev. Try to open and verify
that they actually are ttys by calling isatty().

Signed-off-by: H. Peter Anvin <hpa@zytor.com>